Fix PR structural CI failures and split onboarding flow
